wip: preserve batch 4 contract extraction
